AI-RM-Banking-Chatbot/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
pandas
google-generativeai
drf-yasg
django-redis
